Moved files, deleted superfluous ones, added menu, imported TextMeshPro asset package (#6)
* Renamed folders to fit standard nomencalture

There are a lot of meta files here, but I renamed the folders to Assets, Prefabs, and Materials. I also added a "_Scenes" folder, with hyphen serving to sort it to the top of the list.

* Added basic menu

Added Start, Settings, and Quit button. Will add Load and Continue game buttons and menus as those features are implemented.

* Fixed some conflicts

* Deleted unused script

Deleted extra MainMenu script. Also added Green_Gradient asset to prefabs for text.

* Moving stuff around and deleteing assets incrementally

* Still moving and deleting in increments

* Tiny change

* Finished deleting files I'm sure we don't need