Created End Game function (#12)
* Renamed folders to fit standard nomencalture

There are a lot of meta files here, but I renamed the folders to Assets, Prefabs, and Materials. I also added a "_Scenes" folder, with hyphen serving to sort it to the top of the list.

* Added basic menu

Added Start, Settings, and Quit button. Will add Load and Continue game buttons and menus as those features are implemented.

* Fixed some conflicts

* Deleted unused script

Deleted extra MainMenu script. Also added Green_Gradient asset to prefabs for text.

* Moving stuff around and deleteing assets incrementally

* Still moving and deleting in increments

* Tiny change

* Finished deleting files I'm sure we don't need

* Added pause menu

Pause menu has a Resume, Menu, and Quit button. Added animations folder and a pause menu animation to make it fade in. Also edited the Player_Controller script to prevent the player from firing while the game is paused. Also gave credit to YouTube videos in both menu scripts.

* Added health bar and fixed health UI

Health is now a public static int which can be referenced by other scripts. Health is still changed in the destroy by contact script, but the slider UI is only referenced and updated in the PlayerController script

* Created End game restart function

Modified destroy_by_contact more to reduce how it interacts with player health. Most health logic is included under the playerController now